% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snoqualmie.R
\docType{data}
\name{snoqualmie_data}
\alias{snoqualmie_data}
\title{Daily precipitation at Snoqualmie Falls.}
\format{A data frame with 36 rows, corresponding to years, and 366 variables, corresponding to day within each year:
\describe{
  \item{Day1}{1=measurable precipitation, 0=no measurable precipitation}
  \item{Day2}{1=measurable precipitation, 0=no measurable precipitation}
  ...
}}
\source{
\url{https://www.stat.washington.edu/peter/book.data/set1}
}
\usage{
snoqualmie_data
}
\description{
A dataset containing indicators of rainy days at Snoqualmie Falls.
}
\keyword{datasets}

